Close the CUDA static-analysis lens gap
CI's static job ran in the dev (cpu) image, so CUDA-gated translation units
never entered clang-tidy's compile database and 114 gated findings accumulated
unseen. Fix both halves:

- The findings: 106 implicit widenings (batch loops iterate in std::size_t,
  byte sizes lead with sizeof so multiplication is performed in size_t,
  element-count constructors cast the left operand; ints stay where they feed
  %d printfs and CUDA launcher parameters) and the cublas status switch's
  cloned branches (grouped, enumeration kept as documentation). Behavior
  identical at current parameters; the latent 32-bit-overflow class is gone.
- The seam: the gate's static job and the nightly static leg run in dev-cuda
  (the toolkit puts the TUs in the compile database; clang-tidy parses headers
  only, so no GPU -- the ci overlay drops the device reservation, same as
  build-test). make static prints an advisory when the shell has no toolkit.

Verified: make static in dev-cuda 114 errors -> 0 (exit 0); the changed perf
tests compile; gate 12/12 with the static job running the new lens against
this tree.

+++ src/sim/gpu_compute/fft_analyzer/inc/FFTAnalyzerModel.hpp
@@ -320,7 +320,7 @@ private:
     const float DT = 1.0f / p.sampleRateHz;
 
     for (std::uint32_t ch = 0; ch < p.channelCount; ++ch) {
-      float* chBuf = hInput_.data() + ch * p.samplesPerChannel;
+      float* chBuf = hInput_.data() + static_cast<std::size_t>(ch) * p.samplesPerChannel;
       // Each channel has a distinct dominant frequency + noise
       const float FREQ = 100.0f + static_cast<float>(ch) * 37.5f; // Hz
       const float AMP = 1.0f;
